backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
openai==1.3.7
pytesseract==0.3.10
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
numpy==1.24.3
opencv-python==4.8.1.78
spacy==3.7.2
nltk==3.8.1
